scripts/log_supabase_check.py
#!/usr/bin/env python3
"""
Script to log the results of checking for supabase remnants in the codebase
"""
import sys
from pathlib import Path
from datetime import datetime

# Add parent directories to path
sys.path.append(str(Path(__file__).parent.parent.parent))

# Import logger
from animation_app.utils.logger import Logger

# Create logger instance for web_app
logger = Logger("web_app", log_dir="/app/main/web_app/logs")

# Log the findings
logger.info("Supabase Remnants Check", f"Date: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")

logger.info("Search Results Summary", """
- Main /app directory: No active supabase references found
- Backup folders contain historical supabase references:
  - /app/main/web_app/backup: 19 files with supabase references
  - /app/main/web_app_v1: 78 files with supabase references
  - /app/main/web_app/scripts: 94 files with supabase references (mostly in database/ subfolder)
""")

logger.info("Key Findings", """
1. No active supabase code in main application directories
2. All supabase references are in:
   - Backup folders (marked with _supabase_backup_ timestamps)
   - Deprecated files (prefixed with deprecated_supabase_)
   - Migration scripts (migrate_supabase_to_firebase.py)
   - Historical documentation
""")

logger.info("Migration Status", """
- Successfully migrated from Supabase to Firebase
- Migration completed on 2025-07-16
- All active code now uses Firebase services
- Supabase references only exist in:
  - Backup files for historical reference
  - Migration scripts that handled the transition
  - Deprecated documentation files
""")

logger.success("Supabase Check Complete", """
No active supabase dependencies or code found in production.
All supabase references are properly contained in backup/deprecated/migration files.
The application has been successfully migrated to Firebase.
""")

print("\nSupabase remnants check logged successfully!")
print(f"Log file: {logger.get_log_file_path()}")